refactor(perf): rename file to TestHelpers
