+++ perf/TestHelpers.js
